ANALYZER-1633 - Following Manual Deployment instructions, Export to PDF for geo maps does not work.
the html template used for printing was never updated to work with the amd module loading system/refactorings. also, as a result of this new loading, our loading is a bit slower and our default timeout of 1200 milliseconds is no longer sufficient to wait for the template to completely render with all map tiles.